+++ src/app.py
@@ -2,30 +2,13 @@ import requests
 from flask import Flask, request
 from flask_restful import reqparse, abort, Api, Resource
 
+import resource.vm as VMres
+
 
 app = Flask(__name__)
 api = Api(app)
 
-
-class VMInfo(Resource):
-    """
-    Get vm informations.
-    """
-    def get(self):
-        agents = [
-            {"hostname": '10.8.0.6', "port": 5000},
-        ]
-
-        res = {}
-        for agent in agents:
-            uri = agent["hostname"] + ':' + str(agent["port"]) + "/vm"
-            response = requests.get("http://"+uri).json()
-            res.update({agent["hostname"]: response})
-                        
-        return res
-
-
-api.add_resource(VMInfo, "/vminfo")
+api.add_resource(VMres.VMInfo, "/vminfo")
 
 
 if __name__ == "__main__":
